npm fiddling

+++ wkb.js
@@ -230,7 +230,7 @@ wkb.Polygon.registerParser("WKB", function(instance){
     },
 
     _parse : function(){
-      wkb.Utils.assert(this.data.getUInt32(1) == this.type, "Wrong type for Polygon");
+      wkb.Utils.assert(this.data.getUInt8(1) == this.type, "Wrong type for Polygon");
       var offset = this.byteOffset;
       for(var i = 0; i < this.numRings(); i++){
         var ring = new LineString(new DataView(this.data.buffer, offset));
